Introduce TextPropertyIndex.
