+++ src/kit/app/compareDOMElementPosition.js
@@ -1,0 +1,25 @@
+export const BEFORE = -1
+export const AFTER = 1
+export const PARENT = -2
+export const CHILD = 2
+
+// TODO:this should go into Substance Land
+export function compareDOMElementPosition (a, b) {
+  if (a._isBrowserDOMElement) {
+    let res = a.el.compareDocumentPosition(b.el)
+    if (res & window.Node.DOCUMENT_POSITION_CONTAINS) {
+      return CHILD
+    } else if (res & window.Node.DOCUMENT_POSITION_CONTAINED_BY) {
+      return PARENT
+    } else if (res & window.Node.DOCUMENT_POSITION_PRECEDING) {
+      return AFTER
+    } else if (res & window.Node.DOCUMENT_POSITION_FOLLOWING) {
+      return BEFORE
+    } else {
+      return 0
+    }
+  } else {
+    console.error('FIXME: compareDOMElementPosition() is not implemented for MemoryDOMElement.')
+    return 0
+  }
+}
